refactor the structure of the project, include readme. this is still under development
