+++ .vscode/settings.json
@@ -3,6 +3,6 @@
         "PYTHONPATH": "${workspaceFolder}"
     },
     "editor.codeActionsOnSave": {
-      "source.organizeImports": true
+      "source.organizeImports": "explicit"
     },
 }
